Update and rename Healthy Foods to Healthy Foods List
The paleolithic diet, the diet of our ancestors as a hunter-gatherer society, is what is perfectly safe for humans to eat. Exceptions can include allergens and foods to irritate the gut for certain people. 

I have modified the above list to be vegetarian. Grains, beans, and legumes were removed from the original list.